core, community: move graph vectorstores to community
