+++ libs/community/langchain_community/graph_vectorstores/extractors/link_extractor.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 from typing import Generic, Iterable, Set, TypeVar
 
 from langchain_core._api import beta
-from langchain_core.graph_vectorstores import Link
+from langchain_community.graph_vectorstores import Link
 
 InputT = TypeVar("InputT")
