+++ libs/community/langchain_community/graph_vectorstores/extractors/link_extractor_transformer.py
@@ -3,7 +3,7 @@ from typing import Any, Sequence
 from langchain_core._api import beta
 from langchain_core.documents import Document
 from langchain_core.documents.transformers import BaseDocumentTransformer
-from langchain_core.graph_vectorstores.links import copy_with_links
+from langchain_community.graph_vectorstores.links import copy_with_links
 
 from langchain_community.graph_vectorstores.extractors.link_extractor import (
     LinkExtractor,
