+++ libs/community/tests/integration_tests/graph_vectorstores/extractors/test_gliner_link_extractor.py
@@ -1,5 +1,5 @@
 import pytest
-from langchain_core.graph_vectorstores.links import Link
+from langchain_community.graph_vectorstores.links import Link
 
 from langchain_community.graph_vectorstores.extractors import GLiNERLinkExtractor
