+++ libs/community/tests/unit_tests/graph_vectorstores/extractors/test_html_link_extractor.py
@@ -1,5 +1,5 @@
 import pytest
-from langchain_core.graph_vectorstores import Link
+from langchain_community.graph_vectorstores import Link
 
 from langchain_community.graph_vectorstores.extractors import (
     HtmlInput,
