--- libs/core/langchain_core/graph_vectorstores/__init__.py
@@ -1,15 +1,0 @@
-from langchain_core.graph_vectorstores.base import (
-    GraphVectorStore,
-    GraphVectorStoreRetriever,
-    Node,
-)
-from langchain_core.graph_vectorstores.links import (
-    Link,
-)
-
-__all__ = [
-    "GraphVectorStore",
-    "GraphVectorStoreRetriever",
-    "Node",
-    "Link",
-]
